apps/blender_navigator/backend/requirements.txt
fastapi
uvicorn
google-genai
websockets
python-dotenv
